update workflows
- update runner versions
- update compiler version
- update clang-tidy version
- fix new clang-tidy warnings
- re-enable core guidelines checker
- remove iwyu
- add custom latch implementation to avoid issues with libstdc++ atomic wait/notify implementation on arm64
- fix ARM issues in general

+++ src/utility/cone_tree.cpp
@@ -133,7 +133,7 @@ namespace gapp::detail
             node.radius = findRadius(node_cbegin(node), node_cend(node), node.center);
 
             /* Leaf node. */
-            if (size_t(node.last - node.first) <= MAX_LEAF_ELEMENTS)
+            if (node.last - node.first <= MAX_LEAF_ELEMENTS)
             {
                 node.left = 0;
                 node.right = 0;
